+++ aws_auth.tf
@@ -12,13 +12,21 @@ resource "null_resource" "update_config_map_aws_auth" {
     working_dir = path.module
 
     command = <<EOS
+AWS_AUTH_CONFIG_MAP_APPLIED=false; \
 for i in `seq 1 10`; do \
 echo "${null_resource.update_config_map_aws_auth[0].triggers.kube_config_map_rendered}" > kube_config.yaml & \
 echo "${null_resource.update_config_map_aws_auth[0].triggers.config_map_rendered}" > aws_auth_configmap.yaml & \
-kubectl apply -f aws_auth_configmap.yaml --kubeconfig kube_config.yaml && break || \
+if kubectl apply -f aws_auth_configmap.yaml --kubeconfig kube_config.yaml; then \
+AWS_AUTH_CONFIG_MAP_APPLIED=true; \
+break; \
+fi; \
 sleep 10; \
 done; \
-rm aws_auth_configmap.yaml kube_config.yaml;
+rm aws_auth_configmap.yaml kube_config.yaml; \
+if [ "$AWS_AUTH_CONFIG_MAP_APPLIED" = "false" ]; then \
+echo "Failed all config map apply attempts" >&2; \
+exit 1; \
+fi;
 EOS