+++ CHANGELOG.md
@@ -22,6 +22,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
  - Support map users and roles to multiple groups (by @nauxliu)
  - Fixed errors sometimes happening during destroy due to usage of coalesce() in local.tf (by @petrikero)
  - Removed historical mention of adding caller's IPv4 to cluster security group (by @dpiddockcmp)
+ - Allow `update_config_map_aws_auth` resource apply to fail if all config map apply attempts fail (by @camlow325)
  - Write your awesome change here (by @you)
 
 # History